Refactore Delete item view and functionality
